Fix rand-extension
