+++ rand-extension/Cargo.toml
@@ -1,14 +1,18 @@
 [package]
 name = "rand_extension"
-version = "1.0.0"
+version = "4.3.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2021"
 publish = false
 
 [dependencies]
-ink = { version = "4.3", default-features = false }
+ink = { version = "4.3.0", default-features = false }
+
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2.5", default-features = false, features = ["derive"] }
+scale-info = { version = "2.6", default-features = false, features = ["derive"], optional = true }
+
+[lib]
+path = "lib.rs"
 
 [features]
 default = ["std"]
