--- rand-extension/contract/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "rand_contract"
-version = "4.3.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-publish = false
-
-[dependencies]
-ink = { version = "4.3", default-features = false }
-scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2.5", default-features = false, features = ["derive"] }
-rand_extension = { version = "1.0.0", path = "../rand_extension", default-features = false  }
-
-[lib]
-path = "lib.rs"
-
-[features]
-default = ["std"]
-std = [
-    "ink/std",
-    "scale/std",
-    "scale-info/std",
-    "rand_extension/std"
-]
-ink-as-dependency = []
-
